pyproject.toml
[project]
name = "documentclassification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "asyncpg>=0.30.0",
    "datasets>=3.1.0",
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.115.5",
    "lightning>=2.4.0",
    "lingua-language-detector>=2.0.2",
    "numpy>=2.1.3",
    "pdf2image>=1.17.0",
    "pillow>=11.0.0",
    "psutil>=6.1.0",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.18",
    "rich>=13.9.4",
    "sqlalchemy[asyncio]>=2.0.36",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.46.3",
]
